+++ src/test/java/lotto/manager/LottoManagerTest.java
@@ -1,9 +1,8 @@
-package lotto;
+package lotto.manager;
 
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.UserLotto;
-import lotto.manager.LottoManager;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
